スーパーアイドルタイムの要修正箇所の調整

+++ Assets/Scripts/GameParameter/SuperIdolTimeParamater.cs
@@ -9,11 +9,11 @@ public class SuperIdolTimeParamater : ScriptableObject
     [SerializeField, Header("スーパーアイドルタイムの持続時間")]
     private float _timeEndSuperIdolTime = 15;
     [SerializeField, Header("成功時のスコア量")]
-    private float _successScore = 5000;
+    private int _successScore = 5000;
     /// <summary>ゲージがマックスになる回数</summary>
     public int GaugeCountMax => _gaugeCountMax;
     /// <summary>スーパーアイドルタイムの持続時間</summary>
     public float TimeEndSuperIdolTime => _timeEndSuperIdolTime;
     /// <summary>成功時のスコア量</summary>
-    public float SuccessScore => _successScore;
+    public int SuccessScore => _successScore;
 }
